Create main (splitter) + real-world feedback about program
Create splitter to: load data from file, parse, output in CSV format
Convert weight: int to float (to support .5 kg)
Refactor: Exercise.__repr__

Parser: parse at the level of mini_reps (all children)

+++ parser/test_parser.py
@@ -32,5 +32,5 @@ class TestParser(unittest.TestCase):
 
         self.assertListEqual(result, [Exercise('Row en maquina', [self.serie(i, 41) for i in [15, 8]])])
 
-    def serie(self, repetition: int, weight: int) -> Repetition:
+    def serie(self, repetition: int, weight: float) -> Repetition:
         return {'repetitions': repetition, 'weight': {'amount': weight, 'unit': Units.KILOGRAM}}
